Move service group display labeling from transition-common to frontend
Move `findDistinguishingToken`, `collectServiceNames`, `computeGroupCommonName`,
and `assignCommonNamesForDuplicates` from `PathServiceGrouping.ts` in
transition-common to a new `PathServiceGroupingLabels.ts` in transition-frontend.
These functions produce UI display names for service groups and have no backend
consumers. The domain grouping logic stays in common.
